tg-trade-suite/requirements.txt
python-dotenv
python-telegram-bot==21.*
fastapi>=0.111
uvicorn
asyncpg
